extract chat log
